Modify UserRequest to contain UserID for lookup Modify UserService to error check for bad requests to database Modify UserController to check for bad HTTP requests with missing fields Modify UserController to check for bad responses from UserService and relay to user API requires further testing to confirm it works User Password needs encrypted before being stored in the database

+++ microservices-parent/user-service/src/main/java/com/beefin/services/controller/UserController.java
@@ -5,7 +5,9 @@ import com.beefin.services.dto.UserResponse;
 import com.beefin.services.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 import java.util.concurrent.ExecutionException;
@@ -18,26 +20,73 @@ public class UserController {
     private final UserService userService;
 
     @PostMapping
-    @ResponseStatus(code = HttpStatus.CREATED)
-    public String createUser(@RequestBody UserRequest userRequest) throws ExecutionException, InterruptedException {
-        return userService.createUser(userRequest);
+    public ResponseEntity<String> createUser(@RequestBody UserRequest userRequest) throws ExecutionException, InterruptedException {
+        if ( userRequest.getEmail().isEmpty() || userRequest.getPassword().isEmpty() || userRequest.getName().isEmpty()) {
+            return new ResponseEntity<String>("All fields required", HttpStatus.BAD_REQUEST);
+        }
+
+        HttpStatus code = userService.createUser(userRequest);
+
+        if (code == HttpStatus.CREATED) {
+            return new ResponseEntity<String>("User Created", code);
+        } else if (code == HttpStatus.CONFLICT) {
+            return new ResponseEntity<String>("User already exists", code);
+        } else {
+            return new ResponseEntity<String>("Unexpected error occurred", code);
+        }
     }
 
     @GetMapping
-    @ResponseStatus(code = HttpStatus.ACCEPTED)
+    @ResponseStatus(HttpStatus.OK)
     public List<UserResponse> getAllUsers() throws ExecutionException, InterruptedException {
-        return userService.getAllUsers();
+        List<UserResponse> allUsers = userService.getAllUsers();
+
+        if (allUsers == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Unexpected error occured while getting users");
+        } else {
+            return allUsers;
+        }
     }
 
+    /*
+    @GetMapping
+    @ResponseStatus(HttpStatus.OK)
+    public UserResponse getOneUser(@RequestParam String userID) throws ExecutionException, InterruptedException {
+        UserResponse user = userService.getUser(userID);
+
+        if (user != null) {
+            return user;
+        } else {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "User does not exist or there was a server error");
+        }
+    }
+    */
+
     @PutMapping
-    @ResponseStatus(code = HttpStatus.OK)
-    public String updateUser(@RequestBody UserRequest userRequest) throws ExecutionException, InterruptedException {
-        return userService.update(userRequest);
+    public ResponseEntity<String> updateUser(@RequestBody UserRequest userRequest) throws ExecutionException, InterruptedException {
+        HttpStatus code = userService.update(userRequest);
+
+        if (code == HttpStatus.OK) {
+            return new ResponseEntity<String>("User was updated successfully", code);
+        } else {
+            return new ResponseEntity<String>("Unexpected error occurred while updating", code);
+        }
     }
 
     @DeleteMapping
-    @ResponseStatus(code = HttpStatus.OK)
-    public String deleteUser(@RequestParam String userID) throws ExecutionException, InterruptedException {
-        return userService.deleteUser(userID);
+    public ResponseEntity<String> deleteUser(@RequestParam String userID) throws ExecutionException, InterruptedException {
+        if ( userID.isEmpty()) {
+            return new ResponseEntity<String>("All userID required", HttpStatus.BAD_REQUEST);
+        }
+
+        HttpStatus code = userService.deleteUser(userID);
+
+        if (code == HttpStatus.OK) {
+            return new ResponseEntity<String>("User was deleted successfully", code);
+        } else if (code == HttpStatus.BAD_REQUEST) {
+            return new ResponseEntity<String>("The user to be deleted does not exist", code);
+        } else {
+            return new ResponseEntity<String>("Unexpected error occurred while deleting", code);
+        }
     }
 }
